+++ src/Analysis/ClassListProvider.php
@@ -2,26 +2,10 @@
 
 namespace PhpIntegrator\Analysis;
 
-use PhpIntegrator\Analysis\ClasslikeInfoBuilder;
-use PhpIntegrator\Analysis\ClasslikeInfoBuilderProviderInterface;
-
-use PhpIntegrator\Analysis\Conversion\MethodConverter;
-use PhpIntegrator\Analysis\Conversion\ConstantConverter;
-use PhpIntegrator\Analysis\Conversion\PropertyConverter;
-use PhpIntegrator\Analysis\Conversion\FunctionConverter;
 use PhpIntegrator\Analysis\Conversion\ClasslikeConverter;
-use PhpIntegrator\Analysis\Conversion\ClasslikeConstantConverter;
-
-use PhpIntegrator\Analysis\Relations\TraitUsageResolver;
-use PhpIntegrator\Analysis\Relations\InheritanceResolver;
-use PhpIntegrator\Analysis\Relations\InterfaceImplementationResolver;
-
-use PhpIntegrator\Analysis\Typing\TypeAnalyzer;
 use PhpIntegrator\Analysis\Typing\FileClassListProviderInterface;
 
 use PhpIntegrator\Indexing\IndexDatabase;
-
-use PhpIntegrator\UserInterface\ClasslikeInfoBuilderWhiteHolingProxyProvider;
 
 /**
  * Retrieves a list of available classes.
@@ -29,64 +13,23 @@ use PhpIntegrator\UserInterface\ClasslikeInfoBuilderWhiteHolingProxyProvider;
 class ClassListProvider implements FileClassListProviderInterface
 {
     /**
+     * @var ClasslikeConverter
+     */
+    protected $classlikeConverter;
+
+    /**
      * @var IndexDatabase
      */
     protected $indexDatabase;
 
     /**
-     * @var ClasslikeInfoBuilderWhiteHolingProxyProvider
+     * @param ClasslikeConverter $classlikeConverter
+     * @param IndexDatabase      $indexDatabase
      */
-    protected $storageProxy;
-
-    /**
-     * @var ClasslikeInfoBuilder
-     */
-    protected $classlikeInfoBuilder;
-
-    /**
-     * @param ConstantConverter                     $constantConverter
-     * @param ClasslikeConstantConverter            $classlikeConstantConverter
-     * @param PropertyConverter                     $propertyConverter
-     * @param FunctionConverter                     $functionConverter
-     * @param MethodConverter                       $methodConverter
-     * @param ClasslikeConverter                    $classlikeConverter
-     * @param InheritanceResolver                   $inheritanceResolver
-     * @param InterfaceImplementationResolver       $interfaceImplementationResolver
-     * @param TraitUsageResolver                    $traitUsageResolver
-     * @param ClasslikeInfoBuilderProviderInterface $classlikeInfoBuilderProvider
-     * @param TypeAnalyzer                          $typeAnalyzer
-     * @param IndexDatabase                         $indexDatabase
-     */
-    public function __construct(
-        ConstantConverter $constantConverter,
-        ClasslikeConstantConverter $classlikeConstantConverter,
-        PropertyConverter $propertyConverter,
-        FunctionConverter $functionConverter,
-        MethodConverter $methodConverter,
-        ClasslikeConverter $classlikeConverter,
-        InheritanceResolver $inheritanceResolver,
-        InterfaceImplementationResolver $interfaceImplementationResolver,
-        TraitUsageResolver $traitUsageResolver,
-        ClasslikeInfoBuilderProviderInterface $classlikeInfoBuilderProvider,
-        TypeAnalyzer $typeAnalyzer,
-        IndexDatabase $indexDatabase
-    ) {
+    public function __construct(ClasslikeConverter $classlikeConverter, IndexDatabase $indexDatabase)
+    {
+        $this->classlikeConverter = $classlikeConverter;
         $this->indexDatabase = $indexDatabase;
-        $this->storageProxy = new ClasslikeInfoBuilderWhiteHolingProxyProvider($classlikeInfoBuilderProvider);
-
-        $this->classlikeInfoBuilder = new ClasslikeInfoBuilder(
-            $constantConverter,
-            $classlikeConstantConverter,
-            $propertyConverter,
-            $functionConverter,
-            $methodConverter,
-            $classlikeConverter,
-            $inheritanceResolver,
-            $interfaceImplementationResolver,
-            $traitUsageResolver,
-            $this->storageProxy,
-            $typeAnalyzer
-        );
     }
 
     /**
@@ -117,27 +60,9 @@ class ClassListProvider implements FileClassListProviderInterface
         $result = [];
 
         foreach ($this->indexDatabase->getAllStructuresRawInfo($file) as $element) {
-            $result[$element['fqcn']] = $this->getClassInfoFromRawData($element);
+            $result[$element['fqcn']] = $this->classlikeConverter->convert($element);
         }
 
         return $result;
-    }
-
-    /**
-     * @param array $element
-     *
-     * @return array
-     */
-    protected function getClassInfoFromRawData(array $element): array
-    {
-        // Directly load in the raw information we already have, this avoids performing a database query for each
-        // record.
-        $this->storageProxy->setStructureRawInfo($element);
-
-        $info = $this->classlikeInfoBuilder->getClasslikeInfo($element['name']);
-
-        unset($info['constants'], $info['properties'], $info['methods']);
-
-        return $info;
     }
 }