+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 3.0.0
 * At least PHP 7.1 is now required.
+* PHP 7.1 is now supported (https://github.com/php-integrator/core/issues/82).
+  * It already parsed before, but this involves properly detecting the new scalar types, multiple exception types, ...
 * The class list will now only provide fields directly relevant to the class.
   * Most of the related data, such as methods and constants, were already being filtered out for performance reasons.
   * In order to fetch more information about a class, such as its parents, you now have to manually fetch this using the class info command.