+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * In order to fetch more information about a class, such as its parents, you now have to manually fetch this using the class info command.
 * Linting will now report the fully qualified name of a global function that wasn't found (instead of just the local name).
 * Linting will now report the fully qualified name of a global constant that wasn't found (instead of just the local name).
+* Updated to react/socket 0.5.0.
 
 ## 2.1.5
 * Indexing performance was slightly improved.